requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
matplotlib==3.10.0
reportlab==4.2.5
python-pptx==1.0.2
openpyxl==3.1.5
Pillow==11.1.0
pdfplumber==0.11.4
PyPDF2==3.0.1
python-multipart==0.0.18
sentence-transformers>=3.0.0
torch>=2.0.0
asyncpg>=0.29.0
faker>=22.0.0
numpy>=1.26.0
